fix(wallet): V2 wallet_dormancy_phase column INTEGER not SMALLINT
Hibernate validate-mode required the column type to match the entity's
java.lang.Integer (Types#INTEGER) — SMALLINT mapped to int2 (Types#SMALLINT)
which fails schema validation at startup. All 1591 tests pass after this
correction.

+++ backend/src/main/resources/db/migration/V2__wallet_balance_columns.sql
@@ -12,7 +12,7 @@ ALTER TABLE users
     ADD COLUMN balance_lindens BIGINT NOT NULL DEFAULT 0,
     ADD COLUMN reserved_lindens BIGINT NOT NULL DEFAULT 0,
     ADD COLUMN wallet_dormancy_started_at TIMESTAMPTZ NULL,
-    ADD COLUMN wallet_dormancy_phase SMALLINT NULL,
+    ADD COLUMN wallet_dormancy_phase INTEGER NULL,
     ADD COLUMN wallet_terms_accepted_at TIMESTAMPTZ NULL,
     ADD COLUMN wallet_terms_version VARCHAR(16) NULL;
